backend/requirements.txt
# ─── Django Core ──────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# ─── Auth ─────────────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ─── Realtime (WebSockets) ────────────────────────────────────────────────────
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# ─── Storage (MinIO / S3) ─────────────────────────────────────────────────────
django-storages==1.14.3
boto3==1.34.87

# ─── Email ────────────────────────────────────────────────────────────────────
sendgrid==6.11.0
django-sendgrid-v5==1.2.3

# ─── PDF / Excel Export ───────────────────────────────────────────────────────
reportlab==4.2.0
openpyxl==3.1.2
WeasyPrint==62.1

# ─── Utils ────────────────────────────────────────────────────────────────────
Pillow==10.3.0
python-decouple==3.8
django-uuid-upload-path==1.1.0
python-dateutil==2.9.0.post0
celery==5.4.0
redis==5.0.4

# ─── Dev / Testing ────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
gunicorn==22.0.0
